Move prompt, confirm and sendcastapi to one module each
